🚚(plugins) move custom plugins to a standard location
Sticking to the default, expected standard ansible project tree seems a
good practice. This removes the warning message about our apps lookup
plugin.

Reference: https://docs.ansible.com/ansible/latest/dev_guide/developing_locally.html
