+++ ansible.cfg
@@ -1,6 +1,4 @@
 [defaults]
-filter_plugins = ./plugins/filter:
-lookup_plugins = ./plugins/lookup:
 inventory = ./inventory
 local_tmp = /tmp/ansible
 remote_tmp = /tmp/ansible
