+++ pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-addopts = -v --cov plugins --cov-report term-missing
+addopts = -v --cov lookup_plugins --cov filter_plugins --cov-report term-missing
 python_files =
     test_*.py
     tests.py
